# openapi.json
{"openapi":"3.1.0","info":{"title":"FastAPI","version":"0.1.0"},"paths":{"/auth/token":{"post":{"tags":["auth"],"summary":"Login","operationId":"login_auth_token_post","requestBody":{"content":{"application/x-www-form-urlencoded":{"schema":{"$ref":"#/components/schemas/Body_login_auth_token_post"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Token"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/user/create/":{"post":{"tags":["user"],"summary":"Create a new user","operationId":"create_user_user_create__post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserCreate"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserRead"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/data/{table}/all":{"get":{"tags":["data"],"summary":"Get all rewards","operationId":"get_table_data_data__table__all_get","security":[{"OAuth2PasswordBearer":[]}],"parameters":[{"name":"table","in":"path","required":true,"schema":{"$ref":"#/components/schemas/Tables"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"type":"array","items":{"anyOf":[{"$ref":"#/components/schemas/RewardRead"},{"$ref":"#/components/schemas/ActivityRead"}]},"title":"Response Get Table Data Data  Table  All Get"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/data/reward/add":{"post":{"tags":["data"],"summary":"Create a new reward","operationId":"create_reward_data_reward_add_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/RewardCreate"}}},"required":true},"responses":{"201":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/RewardRead"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}},"security":[{"OAuth2PasswordBearer":[]}]}},"/data/activity/add":{"post":{"tags":["data"],"summary":"Create a new reward","operationId":"create_activity_data_activity_add_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ActivityCreate"}}},"required":true},"responses":{"201":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ActivityCreate"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}},"security":[{"OAuth2PasswordBearer":[]}]}}},"components":{"schemas":{"ActivityCreate":{"properties":{"name":{"type":"string","title":"Name"},"points":{"type":"integer","title":"Points"}},"type":"object","required":["name","points"],"title":"ActivityCreate"},"ActivityRead":{"properties":{"name":{"type":"string","title":"Name"},"points":{"type":"integer","title":"Points"},"id":{"type":"string","format":"uuid","title":"Id"}},"type":"object","required":["name","points","id"],"title":"ActivityRead"},"Body_login_auth_token_post":{"properties":{"grant_type":{"anyOf":[{"type":"string","pattern":"password"},{"type":"null"}],"title":"Grant Type"},"username":{"type":"string","title":"Username"},"password":{"type":"string","title":"Password"},"scope":{"type":"string","title":"Scope","default":""},"client_id":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Client Id"},"client_secret":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Client Secret"}},"type":"object","required":["username","password"],"title":"Body_login_auth_token_post"},"HTTPValidationError":{"properties":{"detail":{"items":{"$ref":"#/components/schemas/ValidationError"},"type":"array","title":"Detail"}},"type":"object","title":"HTTPValidationError"},"RewardCreate":{"properties":{"name":{"type":"string","title":"Name"},"points":{"type":"integer","title":"Points"}},"type":"object","required":["name","points"],"title":"RewardCreate"},"RewardRead":{"properties":{"name":{"type":"string","title":"Name"},"points":{"type":"integer","title":"Points"},"id":{"type":"string","format":"uuid","title":"Id"}},"type":"object","required":["name","points","id"],"title":"RewardRead"},"Tables":{"type":"string","enum":["rewards","activity"],"title":"Tables"},"Token":{"properties":{"access_token":{"type":"string","title":"Access Token"},"token_type":{"type":"string","title":"Token Type"}},"type":"object","required":["access_token","token_type"],"title":"Token"},"UserCreate":{"properties":{"username":{"type":"string","title":"Username"},"password":{"type":"string","title":"Password"}},"type":"object","required":["username","password"],"title":"UserCreate"},"UserRead":{"properties":{"username":{"type":"string","title":"Username"},"id":{"type":"string","format":"uuid","title":"Id"}},"type":"object","required":["username","id"],"title":"UserRead"},"ValidationError":{"properties":{"loc":{"items":{"anyOf":[{"type":"string"},{"type":"integer"}]},"type":"array","title":"Location"},"msg":{"type":"string","title":"Message"},"type":{"type":"string","title":"Error Type"}},"type":"object","required":["loc","msg","type"],"title":"ValidationError"}},"securitySchemes":{"OAuth2PasswordBearer":{"type":"oauth2","flows":{"password":{"scopes":{},"tokenUrl":"auth/token"}}}}}}